Move data classes in OrbitClientData component
To use ProcessData and ModuleData instead of OrbitProcess and
OrbitModule they need to be available in OrbitCore. Thats why the
component OrbitClientData is introduced, which can be used in OrbitGl
and OrbitCore.
