+++ CMakeLists.txt
@@ -146,6 +146,7 @@ endif()
 add_subdirectory(ElfUtils)
 add_subdirectory(OrbitBase)
 add_subdirectory(OrbitCaptureClient)
+add_subdirectory(OrbitClientData)
 add_subdirectory(OrbitClientModel)
 add_subdirectory(OrbitClientProtos)
 add_subdirectory(OrbitClientServices)
