+++ OrbitClientData/ProcessData.cpp
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "ProcessData.h"
+#include "OrbitClientData/ProcessData.h"
 
 #include <memory>
 #include <vector>
