+++ OrbitClientData/include/OrbitClientData/ProcessData.h
@@ -8,9 +8,9 @@
 #include <memory>
 #include <utility>
 
-#include "ModuleData.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Result.h"
+#include "OrbitClientData/ModuleData.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "module.pb.h"
