+++ OrbitCore/Pdb.cpp
@@ -4,7 +4,7 @@
 
 #include "Pdb.h"
 
-#include "FunctionUtils.h"
+#include "OrbitClientData/FunctionUtils.h"
 #include "OrbitProcess.h"
 #include "Path.h"
 #include "ScopeTimer.h"
