+++ OrbitGl/App.cpp
@@ -20,7 +20,6 @@
 #include "CaptureWindow.h"
 #include "Disassembler.h"
 #include "DisassemblyReport.h"
-#include "FunctionUtils.h"
 #include "FunctionsDataView.h"
 #include "GlCanvas.h"
 #include "ImGuiOrbit.h"
@@ -28,6 +27,7 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Result.h"
 #include "OrbitBase/Tracing.h"
+#include "OrbitClientData/FunctionUtils.h"
 #include "OrbitClientModel/CaptureDeserializer.h"
 #include "OrbitClientModel/CaptureSerializer.h"
 #include "Path.h"
