+++ OrbitGl/App.h
@@ -27,13 +27,12 @@
 #include "FunctionsDataView.h"
 #include "LiveFunctionsDataView.h"
 #include "MainThreadExecutor.h"
-#include "ManualInstrumentationManager.h"
-#include "ModuleData.h"
 #include "ModulesDataView.h"
 #include "OrbitBase/Result.h"
 #include "OrbitBase/ThreadPool.h"
 #include "OrbitCaptureClient/CaptureClient.h"
 #include "OrbitCaptureClient/CaptureListener.h"
+#include "OrbitClientData/ModuleData.h"
 #include "OrbitClientServices/CrashManager.h"
 #include "OrbitClientServices/ProcessManager.h"
 #include "OrbitClientServices/TracepointServiceClient.h"
