+++ OrbitGl/SchedulerTrack.cpp
@@ -6,8 +6,8 @@
 
 #include "App.h"
 #include "EventTrack.h"
-#include "FunctionUtils.h"
 #include "GlCanvas.h"
+#include "OrbitClientData/FunctionUtils.h"
 #include "TextBox.h"
 #include "TimeGraph.h"
