+++ OrbitGl/ThreadTrack.cpp
@@ -6,10 +6,10 @@
 
 #include "../Orbit.h"
 #include "App.h"
-#include "FunctionUtils.h"
 #include "GlCanvas.h"
 #include "ManualInstrumentationManager.h"
 #include "OrbitBase/Profiling.h"
+#include "OrbitClientData/FunctionUtils.h"
 #include "TextBox.h"
 #include "TimeGraph.h"
