+++ OrbitGl/TimeGraph.cpp
@@ -15,11 +15,12 @@
 #include "App.h"
 #include "Batcher.h"
 #include "EventTrack.h"
-#include "FunctionUtils.h"
 #include "Geometry.h"
 #include "GlCanvas.h"
 #include "GpuTrack.h"
 #include "GraphTrack.h"
+#include "ManualInstrumentationManager.h"
+#include "OrbitClientData/FunctionUtils.h"
 #include "Params.h"
 #include "PickingManager.h"
 #include "SamplingProfiler.h"
